ポモドーロ学習時間連携完了

+++ src/database/migrations/2022_07_15_113814_create_studies_table.php
@@ -16,7 +16,7 @@ class CreateStudiesTable extends Migration
         Schema::create('studies', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->integer('user_id');
-            $table->integer('hour');
+            $table->float('hour',3,1);
             $table->string('content',255);
             $table->string('language',255);
             $table->string('date',255);
